refactor: move into `@dxup/nuxt`
