+++ packages/nuxt/src/module/named-layout-slots/components/forward.ts
