+++ packages/nuxt/src/module/named-layout-slots/language.ts
